+++ src/resonanceReconstruction/breitWigner/Type.hpp
@@ -4,7 +4,7 @@ protected:
   double target2CompoundWeightRatio;
   EnergyRange energyRange;
   double nucleonNumber;
-
-  #include "resonanceReconstruction/breitWigner/Type/src/neutronWaveNumber.hpp"
+  
   #include "resonanceReconstruction/breitWigner/Type/src/ctor.hpp"  
+  #include "resonanceReconstruction/breitWigner/Type/src/neutronWaveNumber.hpp"
 };
