+++ src/resonanceReconstruction/breitWigner/lvalue/Type/src/ctor.hpp
@@ -1,12 +1,18 @@
 Type( std::vector< resonance::Type >&& resonances,
-      int orbitalAngularMomentum ) :
+      int orbitalAngularMomentum,
+      double targetSpin ) :
   resonances( std::move( resonances ) ),
   weightedQValue( std::nullopt ),
-  orbitalAngularMomentum( orbitalAngularMomentum ){}
+  orbitalAngularMomentum( orbitalAngularMomentum ),
+  statisticalFactorSum( statisticalFactorSummation
+                        ( targetSpin, orbitalAngularMomentum ) ){}
 
 Type( std::vector< resonance::Type >&& resonances,
       int orbitalAngularMomentum,
-      Quantity<ElectronVolts > weightedQValue ) :
+      Quantity< ElectronVolts > weightedQValue,
+      double targetSpin ) :
   resonances( std::move( resonances ) ),
   weightedQValue( weightedQValue ),
-  orbitalAngularMomentum( orbitalAngularMomentum ){} 
+  orbitalAngularMomentum( orbitalAngularMomentum ),
+  statisticalFactorSum( statisticalFactorSummation
+                        ( targetSpin, orbitalAngularMomentum ) ){}
