+++ src/resonanceReconstruction/breitWigner/lvalue/Type/src/penetrationShift.hpp
@@ -1,4 +1,4 @@
 auto penetrationShift( const double channelRatio ) const {
-  return resonanceReconstruction::penetrationShift( this->orbitalAngularMomentum,
-                                                    channelRatio );
+  return resonanceReconstruction::penetrationShift
+    ( this->orbitalAngularMomentum, channelRatio );
 }
