+++ src/resonanceReconstruction/breitWigner/multilevel/Apply/test/resources/Na-22-mt151
@@ -1,3 +1,5 @@
+ 1.102200+4 2.180550+1          0          0          1          01122 2151    1
+ 1.102200+4 1.000000+0          0          0          2          01122 2151    2
  1.000000-5 1.500000+4          1          2          0          01122 2151    3
  3.000000+0 5.700000-1          0          0          1          01122 2151    4
  2.180550+1 0.000000+0          0          0          6          11122 2151    5
