+++ src/resonanceReconstruction/breitWigner/multilevel/Type/src/evaluate1.hpp
@@ -3,9 +3,5 @@ auto evaluate( const Quantity<ElectronVolts> energy,
                Args&&... args ) const {
   const auto radius = this->radius( energy );
   const auto kernel = psiChi( energy, args... );
-  return Parent::evaluate( energy,
-                           kernel,
-                           radius,
-                           radius,
-                           this->radius );
+  return Parent::evaluate( energy, kernel, radius, radius, this->radius );
 }
