+++ src/resonanceReconstruction/breitWigner/multilevel/Type/src/evaluate2.hpp
@@ -4,9 +4,10 @@ auto evaluate( const Quantity<ElectronVolts> energy,
   const auto channelRadius = this->channelRadius( energy );
   const auto scatteringRadius = this->scatteringRadius( energy );
   const auto kernel = psiChi( energy, args... );
+
   return Parent::evaluate( energy,
                            kernel,
                            channelRadius,
                            scatteringRadius,
-                           this->radius );
+                           this->channelRadius );
 }
