+++ src/resonanceReconstruction/breitWigner/multilevel/Type/test/CMakeLists.txt
@@ -1,0 +1,18 @@
+
+add_executable( resonanceReconstruction.breitWigner.multilevel.Type.test Type.test.cpp )
+target_compile_options( resonanceReconstruction.breitWigner.multilevel.Type.test PRIVATE ${${PREFIX}_common_flags}
+$<$<BOOL:${strict}>:${${PREFIX}_strict_flags}>$<$<CONFIG:DEBUG>:
+${${PREFIX}_DEBUG_flags}
+$<$<BOOL:${coverage}>:${${PREFIX}_coverage_flags}>>
+$<$<CONFIG:RELEASE>:
+${${PREFIX}_RELEASE_flags}
+$<$<BOOL:${link_time_optimization}>:${${PREFIX}_link_time_optimization_flags}>
+$<$<BOOL:${nonportable_optimization}>:${${PREFIX}_nonportable_optimization_flags}>>
+
+${CXX_appended_flags} ${resonanceReconstruction_appended_flags} )
+target_link_libraries( resonanceReconstruction.breitWigner.multilevel.Type.test PUBLIC resonanceReconstruction ) 
+file( GLOB resources "resources/*" )
+foreach( resource ${resources})
+    file( COPY "${resource}" DESTINATION "${CMAKE_CURRENT_BINARY_DIR}" )
+endforeach()
+add_test( NAME resonanceReconstruction.breitWigner.multilevel.Type COMMAND resonanceReconstruction.breitWigner.multilevel.Type.test )
