+++ src/resonanceReconstruction/breitWigner/singleLevel/Apply.hpp
@@ -1,9 +1,9 @@
 class Apply {
 protected:
-  template< template BreitWigner >
-  static int nucleonNumber( const BreitWigner& slbw ){ 
+  template< typename BreitWigner >
+  static int nucleonNumber( const BreitWigner& bw ){ 
     constexpr double neutronAMU = 1.00866491588;
-    return std::lround( slbw.LStates().front().AWRI() * neutronAMU );
+    return std::lround( bw.lValues().front().AWRI() * neutronAMU );
   }
 
   #include "resonanceReconstruction/breitWigner/singleLevel/Apply/src/competitiveWidth.hpp"
