+++ src/resonanceReconstruction/breitWigner/singleLevel/Apply/src/call.hpp
@@ -6,7 +6,7 @@ operator()( const ENDF::resolved::SLBW& slbw, Functor&& callback ) const {
     switch( slbw.NAPS() ){
     case 0:
       return callback( build( slbw,
-                              channelRadius( slbw.LStates().front().AWRI() ),
+                              channelRadius( slbw.lValues().front().AWRI() ),
                               radius( slbw.AP() ) ) );
     case 1:
       return callback( build( slbw, radius( slbw.AP() ) ) );
@@ -15,7 +15,7 @@ operator()( const ENDF::resolved::SLBW& slbw, Functor&& callback ) const {
     switch( slbw.NAPS() ){
     case 0:
       return callback( build( slbw,
-                              channelRadius( slbw.LStates().front().AWRI() ),
+                              channelRadius( slbw.lValues().front().AWRI() ),
                               radius( slbw.APE() ) ) );
     case 1:
       return callback( build( slbw, radius( slbw.APE() ) ) );
@@ -24,5 +24,5 @@ operator()( const ENDF::resolved::SLBW& slbw, Functor&& callback ) const {
                               radius( slbw.AP() ),
                               radius( slbw.APE() ) ) );
     }
-  } 
-}
+  }
+}
