+++ src/resonanceReconstruction/breitWigner/singleLevel/Base.hpp
@@ -6,7 +6,7 @@ protected:
   const Derived& derived() const {
     return static_cast< const Derived& >( *this );
   }
-  
+
   #include "resonanceReconstruction/breitWigner/singleLevel/Base/src/lvalues.hpp"
   #include "resonanceReconstruction/breitWigner/singleLevel/Base/src/evaluate.hpp"
