+++ src/resonanceReconstruction/breitWigner/singleLevel/Type/test/Type.test.cpp
@@ -22,7 +22,7 @@ const std::vector< double >& capture();
 
 SCENARIO( "Integration test" ){
   const auto Rh105 = resonances();
-  const auto& isotope = Rh105.isotopes().front();
+  const auto& isotope = Rh105.isotopes.front();
   const auto& energyRange = isotope.energyRanges().front();
   const auto& slbw = std::experimental::get< 1 >( energyRange );
